refactor: modularize PortfolioTab
