+++ src/components/AmountHelpers.tsx
@@ -1,0 +1,68 @@
+import React from "react";
+import styles from "./InputArea.module.css";
+
+interface AmountHelpersProps {
+  showPercentageShortcuts?: boolean;
+  onPercentageClick?: (percentage: number) => void;
+  availableBalance?: React.ReactNode;
+  disabled?: boolean;
+}
+
+const AmountHelpers: React.FC<AmountHelpersProps> = ({
+  showPercentageShortcuts = false,
+  onPercentageClick,
+  availableBalance,
+  disabled = false,
+}) => {
+  if (!showPercentageShortcuts && !availableBalance) {
+    return null;
+  }
+
+  return (
+    <span className={styles.availableBalance}>
+      {showPercentageShortcuts && onPercentageClick && (
+        <span className={styles.percentageShortcuts}>
+          <button
+            className={styles.percentageButton}
+            onClick={() => onPercentageClick(0.25)}
+            type="button"
+            disabled={disabled}
+          >
+            25%
+          </button>
+          {" | "}
+          <button
+            className={styles.percentageButton}
+            onClick={() => onPercentageClick(0.5)}
+            type="button"
+            disabled={disabled}
+          >
+            50%
+          </button>
+          {" | "}
+          <button
+            className={styles.percentageButton}
+            onClick={() => onPercentageClick(0.75)}
+            type="button"
+            disabled={disabled}
+          >
+            75%
+          </button>
+          {" | "}
+          <button
+            className={styles.percentageButton}
+            onClick={() => onPercentageClick(1)}
+            type="button"
+            disabled={disabled}
+          >
+            Max
+          </button>
+          {availableBalance ? " • " : ""}
+        </span>
+      )}
+      {availableBalance && <>Available: {availableBalance}</>}
+    </span>
+  );
+};
+
+export default AmountHelpers;
