+++ src/components/PortfolioTab.tsx
@@ -1,20 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
 import { useState, useEffect } from "react";
-import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useSharedLaserEyes } from "@/context/LaserEyesContext";
-import {
-  fetchPortfolioDataFromApi,
-  QUERY_KEYS,
-  repayLiquidiumLoan,
-  submitRepayPsbt,
-  RepayLiquidiumLoanResponse,
-} from "@/lib/apiClient";
-import { FormattedRuneAmount } from "./FormattedRuneAmount";
-import { FormattedLiquidiumCollateral } from "./FormattedLiquidiumCollateral";
-import { LoanStateEnum, LiquidiumLoanOffer } from "@/types/liquidium";
+import { fetchPortfolioDataFromApi, QUERY_KEYS } from "@/lib/apiClient";
+import { LiquidiumLoanOffer } from "@/types/liquidium";
 import styles from "./PortfolioTab.module.css";
-import Button from "./Button";
+import RunesPortfolioTable from "./RunesPortfolioTable";
+import LiquidiumLoansSection from "./LiquidiumLoansSection";
+import RepayModal from "./RepayModal";
+import { useLiquidiumPortfolio } from "@/hooks/useLiquidiumPortfolio";
 
 type SortField = "name" | "balance" | "value";
 type SortDirection = "asc" | "desc";
@@ -36,16 +30,14 @@ export default function PortfolioTab() {
   );
   const [isLoadingLiquidium, setIsLoadingLiquidium] = useState(false);
   const [liquidiumError, setLiquidiumError] = useState<string | null>(null);
-  const [isRepayingLoanId, setIsRepayingLoanId] = useState<string | null>(null);
 
-  // Add modal state
-  const [repayModal, setRepayModal] = useState<{
-    open: boolean;
-    loan: LiquidiumLoanOffer | null;
-    repayInfo: RepayLiquidiumLoanResponse["data"] | null;
-    loading: boolean;
-    error: string | null;
-  }>({ open: false, loan: null, repayInfo: null, loading: false, error: null });
+  const {
+    isRepayingLoanId,
+    repayModal,
+    handleRepay,
+    handleRepayModalClose,
+    handleRepayModalConfirm,
+  } = useLiquidiumPortfolio({ address, signPsbt });
 
   // Use the new batch API endpoint for runes portfolio
   const {
@@ -244,93 +236,6 @@ export default function PortfolioTab() {
   }, [address]);
 
   // Minimal repay handler
-  const handleRepay = async (loan: LiquidiumLoanOffer) => {
-    setIsRepayingLoanId(loan.id);
-    try {
-      if (!address) throw new Error("Wallet address required");
-      const result = await repayLiquidiumLoan(loan.id, address);
-      if (result.success && result.data) {
-        setRepayModal({
-          open: true,
-          loan,
-          repayInfo: result.data,
-          loading: false,
-          error: null,
-        });
-      } else {
-        setLiquidiumError(result.error || "Failed to prepare repayment");
-      }
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        setLiquidiumError(err.message || "Failed to repay loan");
-      } else {
-        setLiquidiumError("Failed to repay loan");
-      }
-    } finally {
-      setIsRepayingLoanId(null);
-    }
-  };
-
-  const handleRepayModalClose = () => {
-    setRepayModal({
-      open: false,
-      loan: null,
-      repayInfo: null,
-      loading: false,
-      error: null,
-    });
-  };
-
-  const handleRepayModalConfirm = async () => {
-    if (!repayModal.loan || !repayModal.repayInfo || !repayModal.repayInfo.psbt)
-      return;
-    if (!address || !signPsbt) return;
-    setRepayModal((m) => ({ ...m, loading: true, error: null }));
-    try {
-      // Log PSBT and repay info
-      // Convert base64 PSBT to hex if needed
-      const psbtBase64 = repayModal.repayInfo.psbt;
-      // If signPsbt expects hex, convert
-      // LaserEyesContext signPsbt accepts base64, but log both for debugging
-      const psbtHex = Buffer.from(psbtBase64, "base64").toString("hex");
-      // Try both formats if unsure
-      let signResult = await signPsbt(psbtBase64, false, false);
-      if (!signResult?.signedPsbtBase64 && psbtHex) {
-        signResult = await signPsbt(psbtHex, false, false);
-      }
-      const signedPsbt =
-        signResult?.signedPsbtBase64 || signResult?.signedPsbtHex;
-      if (!signedPsbt) throw new Error("Wallet did not return a signed PSBT");
-      const submitResult = await submitRepayPsbt(
-        repayModal.loan.id,
-        signedPsbt,
-        address,
-      );
-      if (submitResult.success && submitResult.data) {
-        setRepayModal({ ...repayModal, loading: false, error: null });
-        alert(
-          `Repayment submitted! TxID: ${submitResult.data.repayment_transaction_id}`,
-        );
-        handleRepayModalClose();
-        fetchLiquidiumLoans(); // Refresh loans
-      } else {
-        setRepayModal((m) => ({
-          ...m,
-          loading: false,
-          error: submitResult.error || "Failed to submit repayment",
-        }));
-      }
-    } catch (err: unknown) {
-      setRepayModal((m) => ({
-        ...m,
-        loading: false,
-        error:
-          err instanceof Error
-            ? err.message
-            : "Failed to sign or submit repayment",
-      }));
-    }
-  };
 
   // Log repayModal state when opening modal
   useEffect(() => {
@@ -429,37 +334,6 @@ export default function PortfolioTab() {
   );
 
   // Helper function to format loan status
-  const getLoanStatusClass = (status: LoanStateEnum): string => {
-    switch (status) {
-      case "ACTIVE":
-        return styles.statusActive;
-      case "ACTIVATING":
-        return styles.statusActivating;
-      case "REPAYING":
-        return styles.statusRepaying;
-      case "DEFAULTED":
-        return styles.statusDefaulted;
-      case "LIQUIDATED":
-        return styles.statusLiquidated;
-      case "REPAID":
-        return styles.statusRepaid;
-      default:
-        return "";
-    }
-  };
-
-  // Helper function to format date and time on separate lines
-  const formatDate = (dateString: string): React.ReactNode => {
-    const date = new Date(dateString);
-    return (
-      <div className={styles.dateTimeContainer}>
-        <div className={styles.dateDisplay}>{date.toLocaleDateString()}</div>
-        <div className={styles.timeDisplay}>
-          {date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" })}
-        </div>
-      </div>
-    );
-  };
 
   // Helper function to format sats to BTC
   const formatSatsToBtc = (sats: number): string => (sats / 1e8).toFixed(8);
@@ -467,109 +341,15 @@ export default function PortfolioTab() {
   return (
     <div className={styles.container}>
       {/* Runes Portfolio Section */}
-      <div className={styles.listContainer}>
-        <div className={`${styles.listHeader} ${styles.grid4col}`}>
-          <div
-            className="sortable"
-            style={{ fontWeight: "bold" }}
-            onClick={() => handleSort("name")}
-          >
-            Rune Name
-            {sortField === "name" && (
-              <span className={styles.sortArrow}>
-                {sortDirection === "asc" ? "↑" : "↓"}
-              </span>
-            )}
-          </div>
-          <div
-            className="sortable"
-            style={{ fontWeight: "bold" }}
-            onClick={() => handleSort("balance")}
-          >
-            Balance
-            {sortField === "balance" && (
-              <span className={styles.sortArrow}>
-                {sortDirection === "asc" ? "↑" : "↓"}
-              </span>
-            )}
-          </div>
-          <div
-            className="sortable"
-            style={{ fontWeight: "bold" }}
-            onClick={() => handleSort("value")}
-          >
-            Value (USD)
-            {sortField === "value" && (
-              <span className={styles.sortArrow}>
-                {sortDirection === "asc" ? "↑" : "↓"}
-              </span>
-            )}
-          </div>
-          <div style={{ fontWeight: "bold" }}>Action</div>
-        </div>
-        <div className={styles.listContent}>
-          {sortedBalances.map((rune) => {
-            const marketInfo = portfolioData.marketData?.[rune.name];
-            const usdValue = marketInfo?.price_in_usd
-              ? rune.usdValue.toFixed(2)
-              : "0.00";
-
-            return (
-              <div
-                key={rune.name}
-                className={`${styles.listItem} ${styles.grid4col}`}
-              >
-                <div className={styles.runeName}>
-                  <div className={styles.runeNameContent}>
-                    {rune.imageURI && (
-                      <Image
-                        src={rune.imageURI}
-                        alt=""
-                        className={styles.runeImage}
-                        width={24}
-                        height={24}
-                        onError={(e) => {
-                          const target = e.target as HTMLImageElement;
-                          if (target) {
-                            target.style.display = "none";
-                          }
-                        }}
-                      />
-                    )}
-                    <div className={styles.runeNameText}>
-                      <div className={styles.runeFullName}>
-                        {rune.formattedName}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div className={styles.runeBalance}>
-                  <FormattedRuneAmount
-                    runeName={rune.name}
-                    rawAmount={rune.balance}
-                  />
-                </div>
-                <div className={styles.runeValue}>
-                  {!portfolioData ? "..." : `$${usdValue}`}
-                </div>
-                <Button onClick={() => handleSwap(rune.name)}>Swap</Button>
-              </div>
-            );
-          })}
-        </div>
-        <div className={`${styles.portfolioTotals} ${styles.grid4col}`}>
-          <div>Portfolio Total:</div>
-          <div>≈ {totalBtcValue.toFixed(8)} BTC</div>
-          <div>
-            $
-            {totalUsdValue.toLocaleString(undefined, {
-              minimumFractionDigits: 2,
-              maximumFractionDigits: 2,
-            })}
-          </div>
-          <div></div>
-        </div>
-      </div>
+      <RunesPortfolioTable
+        balances={sortedBalances}
+        totalBtcValue={totalBtcValue}
+        totalUsdValue={totalUsdValue}
+        sortField={sortField}
+        sortDirection={sortDirection}
+        onSort={handleSort}
+        onSwap={handleSwap}
+      />
 
       {/* Liquidium Loans Section */}
       <div className={styles.sectionDivider}>
@@ -580,156 +360,36 @@ export default function PortfolioTab() {
         </div>
       </div>
 
-      <div className={styles.liquidiumContainer}>
-        <div className={`${styles.liquidiumHeader} ${styles.grid6col}`}>
-          <div style={{ fontWeight: "bold" }}>Collateral</div>
-          <div style={{ fontWeight: "bold" }}>Principal</div>
-          <div style={{ fontWeight: "bold" }}>Status</div>
-          <div style={{ fontWeight: "bold" }}>Due Date</div>
-          <div style={{ fontWeight: "bold" }}>Repayment</div>
-          <div style={{ fontWeight: "bold" }}>Action</div>
-        </div>
-        <div className={styles.listContent}>
-          {isCheckingAuth ? (
-            <div>Checking Liquidium connection...</div>
-          ) : !liquidiumAuthenticated ? (
-            <div className={styles.liquidiumAuth}>
-              <Button
-                onClick={handleLiquidiumAuth}
-                disabled={isAuthenticating || !address || !signMessage}
-              >
-                {isAuthenticating
-                  ? "Authenticating..."
-                  : "Connect to Liquidium"}
-              </Button>
-              {authError && (
-                <div className="errorText">
-                  {typeof authError === "string"
-                    ? authError
-                    : JSON.stringify(authError)}
-                </div>
-              )}
-            </div>
-          ) : isLoadingLiquidium ? (
-            <div>Loading Liquidium loans...</div>
-          ) : liquidiumError ? (
-            <div className="errorText">
-              {typeof liquidiumError === "string"
-                ? liquidiumError
-                : JSON.stringify(liquidiumError)}
-            </div>
-          ) : !liquidiumLoans.length ? (
-            <div>No Liquidium loans found</div>
-          ) : (
-            liquidiumLoans.map((loan: LiquidiumLoanOffer) => (
-              <div
-                key={loan.id}
-                className={`${styles.liquidiumItem} ${styles.grid6col}`}
-              >
-                <div>
-                  <FormattedLiquidiumCollateral
-                    runeId={loan.collateral_details.rune_id}
-                    runeAmount={loan.collateral_details.rune_amount}
-                    runeDivisibility={loan.collateral_details.rune_divisibility}
-                  />
-                </div>
-                <div className={styles.btcValueContainer}>
-                  <div className={styles.btcAmount}>
-                    {formatSatsToBtc(loan.loan_details.principal_amount_sats)}
-                  </div>
-                  <div className={styles.btcLabel}>BTC</div>
-                </div>
-                <div className={styles.statusContainer}>
-                  <span
-                    className={`${styles.loanStatus} ${getLoanStatusClass(loan.loan_details.state)}`}
-                  >
-                    {loan.loan_details.state}
-                  </span>
-                </div>
-                <div>{formatDate(loan.loan_details.loan_term_end_date)}</div>
-                <div className={styles.btcValueContainer}>
-                  <div className={styles.btcAmount}>
-                    {loan.loan_details.total_repayment_sats
-                      ? formatSatsToBtc(loan.loan_details.total_repayment_sats)
-                      : formatSatsToBtc(
-                          loan.loan_details.principal_amount_sats *
-                            (1 + loan.loan_details.discount.discount_rate),
-                        )}
-                  </div>
-                  <div className={styles.btcLabel}>BTC</div>
-                </div>
-                <div>
-                  {loan.loan_details.state === "ACTIVE" && (
-                    // Temporarily disable repay button due to lasereyes issue
-                    <Button
-                      onClick={() => handleRepay(loan)}
-                      disabled={true}
-                      className={styles.repayButtonDisabled}
-                      //disabled={isRepayingLoanId === loan.id}
-                    >
-                      {isRepayingLoanId === loan.id ? "Repaying..." : "Repay"}
-                    </Button>
-                  )}
-                  {loan.loan_details.state === "ACTIVATING" && (
-                    <Button disabled={true}>Activating...</Button>
-                  )}
-                  {loan.loan_details.state === "REPAYING" && (
-                    <Button disabled={true}>Processing...</Button>
-                  )}
-                </div>
-              </div>
-            ))
-          )}
-        </div>
-      </div>
+      <LiquidiumLoansSection
+        loans={liquidiumLoans}
+        isCheckingAuth={isCheckingAuth}
+        liquidiumAuthenticated={liquidiumAuthenticated}
+        isAuthenticating={isAuthenticating}
+        authError={authError}
+        isLoadingLiquidium={isLoadingLiquidium}
+        liquidiumError={liquidiumError}
+        isRepayingLoanId={isRepayingLoanId}
+        onAuth={handleLiquidiumAuth}
+        onRepay={handleRepay}
+      />
 
-      {/* Repayment Confirmation Modal */}
-      {repayModal.open && (
-        <div className={styles.repayModalOverlay}>
-          <div className={styles.repayModalWindow}>
-            <h3 className="heading">Confirm Repayment</h3>
-            <div>
-              Repayment Amount:{" "}
-              <b>
-                {repayModal.loan
-                  ? `${formatSatsToBtc(
-                      repayModal.loan.loan_details.total_repayment_sats ??
-                        repayModal.loan.loan_details.principal_amount_sats *
-                          (1 +
-                            repayModal.loan.loan_details.discount
-                              .discount_rate),
-                    )} BTC`
-                  : "..."}
-              </b>
-            </div>
-            <div
-              className="smallText"
-              style={{ margin: "8px 0", wordBreak: "break-all" }}
-            >
-              PSBT: <code>{repayModal.repayInfo?.psbt?.slice(0, 32)}...</code>
-            </div>
-            {repayModal.error && (
-              <div className="errorText" style={{ margin: "8px 0" }}>
-                {repayModal.error}
-              </div>
-            )}
-            <div style={{ display: "flex", gap: 8, marginTop: 16 }}>
-              <Button
-                onClick={handleRepayModalClose}
-                disabled={repayModal.loading}
-              >
-                Cancel
-              </Button>
-              <Button
-                onClick={handleRepayModalConfirm}
-                disabled={repayModal.loading}
-              >
-                {repayModal.loading ? "Processing..." : "Sign & Repay"}
-              </Button>
-            </div>
-          </div>
-        </div>
-      )}
+      <RepayModal
+        open={repayModal.open}
+        repayAmount={
+          repayModal.loan
+            ? `${formatSatsToBtc(
+                repayModal.loan.loan_details.total_repayment_sats ??
+                  repayModal.loan.loan_details.principal_amount_sats *
+                    (1 + repayModal.loan.loan_details.discount.discount_rate),
+              )} BTC`
+            : "..."
+        }
+        psbtPreview={repayModal.repayInfo?.psbt?.slice(0, 32) || ""}
+        loading={repayModal.loading}
+        error={repayModal.error}
+        onCancel={handleRepayModalClose}
+        onConfirm={handleRepayModalConfirm}
+      />
     </div>
   );
 }
